Ajustes nas classes.

+++ Trabalho/src/entidade/Negociavel.java
@@ -9,6 +9,6 @@ public interface Negociavel {
 
 	public void adicionaCliente(String nome, String cnpj);
 	public void adicionaImpressoras(String numSerie, String marca,String modelo);
-	public void setFranquiaPaginas(Integer franquiaPaginas);	
+	public void setFranquiaPaginas(Double franquiaPaginas);	
 
 }
